[3242][FIX] mrp_subcontracting_analytic: switch to extend _subcontracted_produce()
Before this commit, _prepare_subcontract_mo_vals() was extended in a seemingly
standard manner to pass the value of analytic distribution from the originating
receipt move to the subcontracted production, however that would somehow generate
duplicated stock moves (for both components and produced product) which was
incorrect.

This commit fixes that problematic behavior.