[4346][FIX] mrp_subcontracting_analytic: revert 1345a6d
We need to include analytic_distribution in the vals used to create MO,
instead of post-processing the update, because there is issue with passing
down the analytic distribution to procurements when procurement_mto_analytic
is used.

The issue reported in the message of 1345a6d does not seem to reproduce with
this change.