[5295][IMP] stock_quant_manual_assign: Improve assgn_quants() performance (#184)
[5295][IMP] stock_quant_manual_assign: Improve assign_quants() performance

Previously, when there were hundreds of quants assigned to a stock move, it
would take a very long time for assign_quants() to complete, as the system
would try to reassign all selected quants.

This commit mitigates the issue by identifying only the necessary changes —
specifically, the move lines to be unlinked and the quant lines to be
assigned — so that updates are made only to those records.